Union hack work before feedback
I think Egyptian flag is too complex so I think it should simple, that is the reason I have designed it. I think bird should simple.
[ Egyptian flag. The three colours of the Egyptian flag, along with its national emblem, represent the nation's modern history, from monarchy rule to British occupation to the end of oppression. The flag's red stripe represents the period prior to the 1952 revolution when King Farouk ruled the nation. The white stripe symbolises the revolution, and black stripe refers to the beginning of the modern form of government.
The flag also includes Egypt's Coat of Arms also known as the Eagle of Saladin, a golden eagle with its head turned toward its right. The eagle represents the Ayyubid sultan who ruled Egypt throughout the 12th century. The emblem also contains a shield and scroll bearing Egypt's name in Arabic. The shield is superimposed over the eagle, while the scroll is placed underneath both. The Eagle of Saladin replaced a golden hawk in 1984