After the feedback
This flag is for UK. [The Union Flag, or Union Jack, is the national flag of the United Kingdom. It is so called because it combines the crosses of the three countries united under one Sovereign - the kingdoms of England and Wales, of Scotland and of Ireland (although since 1921 only Northern Ireland has been part of the United Kingdom].
Here I have used rectangle because it has four coner, each coner represent different states which is England, Wales, Scotland and Northern Ireland.
1 Star represent United as one.

The Egyptian take Islam seriously, green color symbolize Islam that is the reason I used green.
Left hand side I have put shape of pyramids which is to show past Egypt rich history.